// generated/orval/fetch/google-health-api/models/nutritionLogMealType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */

/**
 * Optional. The meal category. One of `BREAKFAST`, `LUNCH`, `DINNER`, or `SNACK`.
 */
export type NutritionLogMealType = typeof NutritionLogMealType[keyof typeof NutritionLogMealType];


export const NutritionLogMealType = {
  MEAL_TYPE_UNSPECIFIED: 'MEAL_TYPE_UNSPECIFIED',
  BEFORE_BREAKFAST: 'BEFORE_BREAKFAST',
  BREAKFAST: 'BREAKFAST',
  BEFORE_LUNCH: 'BEFORE_LUNCH',
  LUNCH: 'LUNCH',
  BEFORE_DINNER: 'BEFORE_DINNER',
  DINNER: 'DINNER',
  AFTER_DINNER: 'AFTER_DINNER',
  SNACK: 'SNACK',
  ANYTIME: 'ANYTIME',
} as const;